"Introduce mood-based projectile scaling, add MoodyWindChargeEntity and custom renderer, and enhance Happy Ghast shooting mechanics with mood multipliers. Refactor projectile damage/radius handling.

+++ src/main/java/ca/corruptdata/moodyghasts/entity/projectile/AbstractIceChargeEntity.java
@@ -41,7 +41,7 @@ import java.util.WeakHashMap;
 public abstract class AbstractIceChargeEntity extends AbstractHurtingProjectile implements ItemSupplier {
 
     protected abstract int getEffectRadius();
-    protected abstract int getDamage();
+    protected abstract float getDamage();
     private final Set<BlockPos> recentlyConverted = Collections.newSetFromMap(new WeakHashMap<>());
     private static final int CONVERSION_TIMEOUT = 2; // ticks
     private static final ResourceKey<DamageType> ICECHARGE_DAMAGE =
